Extract interfaces ot separate files
