+++ logging/src/app/search-form/data/log-stream.ts
@@ -1,0 +1,5 @@
+export interface ILogStream {
+  availableLabels: string[];
+  entries: any[];
+  labels: string;
+}
