+++ logging/src/app/search-form/data/pod-query.ts
@@ -1,0 +1,9 @@
+export interface IPod {
+  name: string;
+  labels: any;
+}
+export interface IPodQueryResponse {
+  data: {
+    pods: IPod[];
+  };
+}
